pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mingli-mcp"
version = "1.0.0"
description = "命理MCP服务 - 支持紫微斗数、八字等多种命理系统"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "spyfree", email = "srlixin@gmail.com"}
]
keywords = ["mcp", "ziwei", "bazi", "fortune", "astrology"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "py-iztro>=0.1.5",
    "lunar_python>=1.4.7",
    "bidict>=0.23.0",
    "colorama>=0.4.6",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/spyfree/mingli-mcp"
Repository = "https://github.com/spyfree/mingli-mcp"
Documentation = "https://github.com/spyfree/mingli-mcp/blob/main/README.md"
Issues = "https://github.com/spyfree/mingli-mcp/issues"

[project.scripts]
mingli-mcp = "mingli_mcp:main"

[tool.setuptools]
packages = ["systems", "transports", "utils", "core"]

[tool.setuptools.package-data]
systems = ["bazi/lib/*.py"]
